Moved `copyPackageJson` and `install` (as `npmInstall`) from `scaffold.js` to `util.js`
